+++ test/tweetnacl/crypto_box_test.rb
@@ -9,4 +9,24 @@ class CryptoBoxTest < MiniTest::Test
     TweetNaCl::CryptoBox.new
     assert :did_not_raise, "A CryptoBox object can be created without arguments"
   end
+
+  def test_can_box_a_message_with_a_nonce
+    require 'pp'
+    message = "hello world"
+    expected_cipher = "0000000000000000FBC937C3F136E09FA8A45C58C15E801394F5BB74CE8D538FE3D726"
+    nonce = ValidNonce
+    nonce = "*" * 24
+
+    sender , receiver = TweetNaCl::KeyPair.new, TweetNaCl::KeyPair.new
+
+    encryption_keypair = TweetNaCl::KeyPair.new([receiver.public_key, sender.secret_key])
+    encrypted_box      = TweetNaCl::CryptoBox.new(encryption_keypair)
+    encrypted_box.close(message, nonce)
+
+    decryption_keypair = TweetNaCl::KeyPair.new([sender.public_key, receiver.secret_key])
+    decryption_box     = TweetNaCl::CryptoBox.new(decryption_keypair)
+    decryption_box.open(encrypted_box, nonce)
+
+    assert_equal message, decryption_box.message
+  end
 end